multi_agents/requirements.txt
json5
langgraph
langgraph-cli
loguru
python-dotenv
weasyprint; sys_platform != 'win32'
